front-end changes fix
