+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/view/EditInstrumentController.java
@@ -34,13 +34,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Instrument;
-import uk.ac.bbsrc.tgac.miso.core.data.InstrumentPosition;
 import uk.ac.bbsrc.tgac.miso.core.data.ServiceRecord;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.InstrumentImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.type.InstrumentType;
@@ -92,34 +90,6 @@ public class EditInstrumentController {
     return setupForm(instrument, model);
   }
 
-  @GetMapping(value = "/{instrumentId}/serviceRecord/new/{recordId}")
-  public ModelAndView newServiceRecord(@PathVariable(value = "instrumentId") Long instrumentId,
-      @PathVariable(value = "recordId") Long recordId, ModelMap model) throws IOException {
-    Instrument instrument = instrumentService.get(instrumentId);
-    if (instrument == null) {
-      throw new NotFoundException("No instrument found for ID " + instrumentId.toString());
-    }
-    ServiceRecord record = serviceRecordService.get(recordId);
-    instrumentService.addServiceRecord(record, instrument);
-    return showPage(record, instrument, model);
-  }
-
-  @GetMapping(value = "/{instrumentId}/serviceRecord/edit/{recordId}")
-  public ModelAndView editServiceRecord(@PathVariable(value = "instrumentId") Long instrumentId,
-      @PathVariable(value = "recordId") Long recordId, ModelMap model) throws IOException {
-    Instrument instrument = instrumentService.get(instrumentId);
-    if (instrument == null) {
-      throw new NotFoundException("No instrument found for ID " + instrumentId.toString());
-    }
-    ServiceRecord record = serviceRecordService.get(recordId);
-    if (record == null) {
-      throw new NotFoundException("No record found for ID " + recordId.toString());
-    }
-
-    instrumentService.updateServiceRecord(record, instrument);
-    return showPage(record, instrument, model);
-  }
-
   private ModelAndView setupForm(Instrument instrument, ModelMap model) throws IOException {
     InstrumentDto instrumentDto = Dtos.asDto(instrument);
 
@@ -155,24 +125,4 @@ public class EditInstrumentController {
 
     return new ModelAndView("/WEB-INF/pages/editInstrument.jsp", model);
   }
-
-  public ModelAndView showPage(ServiceRecord record, Instrument instrument, ModelMap model)
-      throws JsonProcessingException {
-    if (!record.isSaved()) {
-      model.put("title", "New Service Record");
-    } else {
-      model.put("title", "Service Record " + record.getId());
-    }
-    model.put("serviceRecord", record);
-    model.put("serviceRecordDto", mapper.writeValueAsString(Dtos.asDto(record)));
-    ArrayNode positions = mapper.createArrayNode();
-    for (InstrumentPosition pos : instrument.getInstrumentModel().getPositions()) {
-      ObjectNode dto = positions.addObject();
-      dto.put("id", pos.getId());
-      dto.put("alias", pos.getAlias());
-    }
-    model.put("instrumentPositions", mapper.writeValueAsString(positions));
-    return new ModelAndView("/WEB-INF/pages/editServiceRecord.jsp", model);
-  }
-
 }
