+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/view/EditServiceRecordController.java
@@ -20,17 +20,31 @@
 
 package uk.ac.bbsrc.tgac.miso.webapp.controller.view;
 
+import java.io.IOException;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.acls.model.NotFoundException;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Instrument;
+import uk.ac.bbsrc.tgac.miso.core.data.InstrumentPosition;
+import uk.ac.bbsrc.tgac.miso.core.data.ServiceRecord;
 import uk.ac.bbsrc.tgac.miso.core.service.InstrumentService;
 import uk.ac.bbsrc.tgac.miso.core.service.ServiceRecordService;
+import uk.ac.bbsrc.tgac.miso.dto.Dtos;
 
 @Controller
-@RequestMapping("/instrument/servicerecord")
+@RequestMapping("/instrument/{instrumentId}/servicerecord")
 public class EditServiceRecordController {
 
   @Autowired
@@ -40,47 +54,61 @@ public class EditServiceRecordController {
   @Autowired
   private ObjectMapper mapper;
 
-  // @GetMapping(value = "/{recordId}")
-  // public ModelAndView viewServiceRecord(@PathVariable(value = "recordId") Long recordId, ModelMap
-  // model)
-  // throws IOException {
-  // ServiceRecord record = serviceRecordService.get(recordId);
-  // if (record == null) {
-  // throw new NotFoundException("No service found for ID " + recordId.toString());
-  // }
-  // return showPage(record, model);
-  // }
+  @GetMapping(value = "/{recordId}")
+  public ModelAndView viewServiceRecord(@PathVariable(value = "recordId") Long recordId, Long instrumentId,
+      ModelMap model)
+      throws IOException {
+    ServiceRecord record = serviceRecordService.get(recordId);
+    Instrument instrument = instrumentService.get(instrumentId);
+    if (record == null) {
+      throw new NotFoundException("No service found for ID " + recordId.toString());
+    }
+    return showPage(record, instrument, model);
+  }
 
-  // @GetMapping(value = "/new/{instrumentId}")
-  // public ModelAndView newServiceRecord(@PathVariable(value = "instrumentId") Long instrumentId,
-  // ModelMap model)
-  // throws IOException {
-  // Instrument instrument = instrumentService.get(instrumentId);
-  // if (instrument == null) {
-  // throw new NotFoundException("No instrument found for ID " + instrumentId.toString());
-  // }
-  // ServiceRecord record = new ServiceRecord();
-  // record.setInstrument(instrument);
-  // return showPage(record, model);
-  // }
+  @GetMapping(value = "/new/{recordId}")
+  public ModelAndView newServiceRecord(@PathVariable(value = "recordId") Long recordId,
+      Long instrumentId, ModelMap model)
+      throws IOException {
+    Instrument instrument = instrumentService.get(instrumentId);
+    if (instrument == null) {
+      throw new NotFoundException("No instrument found for ID " + instrumentId.toString());
+    }
+    ServiceRecord record = new ServiceRecord();
+    instrumentService.addServiceRecord(record, instrument);
+    return showPage(record, instrument, model);
+  }
 
-  // public ModelAndView showPage(ServiceRecord record, ModelMap model) throws JsonProcessingException
-  // {
-  // if (!record.isSaved()) {
-  // model.put("title", "New Service Record");
-  // } else {
-  // model.put("title", "Service Record " + record.getId());
-  // }
-  // model.put("serviceRecord", record);
-  // model.put("serviceRecordDto", mapper.writeValueAsString(Dtos.asDto(record)));
-  // ArrayNode positions = mapper.createArrayNode();
-  // for (InstrumentPosition pos : record.getInstrument().getInstrumentModel().getPositions()) {
-  // ObjectNode dto = positions.addObject();
-  // dto.put("id", pos.getId());
-  // dto.put("alias", pos.getAlias());
-  // }
-  // model.put("instrumentPositions", mapper.writeValueAsString(positions));
-  // return new ModelAndView("/WEB-INF/pages/editServiceRecord.jsp", model);
-  // }
+  @GetMapping(value = "/edit/{recordId}")
+  public ModelAndView editServiceRecord(@PathVariable(value = "recordId") Long recordId,
+      Long instrumentId, ModelMap model) throws IOException {
+    Instrument instrument = instrumentService.get(instrumentId);
+    if (instrument == null) {
+      throw new NotFoundException("No instrument found for ID " + instrumentId.toString());
+    }
+    ServiceRecord record = serviceRecordService.get(recordId);
+    instrumentService.updateServiceRecord(record, instrument);
+    return showPage(record, instrument, model);
+  }
+
+
+  public ModelAndView showPage(ServiceRecord record, Instrument instrument, ModelMap model)
+      throws JsonProcessingException {
+    if (!record.isSaved()) {
+      model.put("title", "New Service Record");
+    } else {
+      model.put("title", "Service Record " + record.getId());
+    }
+    model.put("serviceRecord", record);
+    model.put("serviceRecordDto", mapper.writeValueAsString(Dtos.asDto(record)));
+    ArrayNode positions = mapper.createArrayNode();
+    for (InstrumentPosition pos : instrument.getInstrumentModel().getPositions()) {
+      ObjectNode dto = positions.addObject();
+      dto.put("id", pos.getId());
+      dto.put("alias", pos.getAlias());
+    }
+    model.put("instrumentPositions", mapper.writeValueAsString(positions));
+    return new ModelAndView("/WEB-INF/pages/editServiceRecord.jsp", model);
+  }
 
 }
